pyproject.toml
[project]
name = "llm-quantization"
version = "0.1.0"
description = "A simple project testing model quantization with LLMs"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "huggingface>=0.0.1",
    "hydra-core>=1.3.2",
    "omegaconf>=2.3.0",
    "transformers>=4.57.1",
    "datasets>=2.0.0",
    "peft>=0.8.0",
    "accelerate>=0.30.0",
    "bitsandbytes>=0.43.0",
]

[dependency-groups]
dev = [
    "ruff>=0.14.5",
]
